move relocatinn code to relocate
