+++ lib/spack/spack/relocate.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
 import collections
 import itertools
+import ctypes
 import os
 import re
 import struct
@@ -14,10 +15,10 @@ import spack.vendor.macholib.MachO
 
 import spack.spec
 import spack.store
+import spack.user_environment
 import spack.util.elf as elf
 import spack.util.executable as executable
 import spack.util.filesystem as fs
-import spack.util.filesystem as sfs
 import spack.util.lang
 from spack.util import elf, executable, tty
 from spack.util.environment import EnvironmentModifications
@@ -25,6 +26,9 @@ from spack.util.filesystem import readlink, symlink
 from spack.util.lang import memoized
 
 from .relocate_text import BinaryFilePrefixReplacer, PrefixToPrefix, TextFilePrefixReplacer
+
+if sys.platform == "win32":
+    import ctypes.wintypes
 
 
 @memoized
@@ -39,22 +43,262 @@ def _patchelf() -> Optional[executable.Executable]:
         return spack.bootstrap.ensure_patchelf_in_path_or_raise()
 
 
-@memoized
-def _msvc_relocate() -> Optional[executable.Executable]:
-    """Returns the full path to the Windows compiler wrapper if installed"""
-
-    # TODO: Should we query the Spack store instead?
-    # is it worth bootstrapping here?
-    import spack.bootstrap
-
-    if sys.platform != "win32":
-        raise RuntimeError("Cannot perform MSVC based relocation on non Windows platform")
-    with spack.bootstrap.ensure_bootstrap_configuration():
-        return spack.bootstrap.ensure_msvc_relocate_or_raise()  # type: ignore
-
-
 def _decode_macho_data(bytestring):
     return bytestring.rstrip(b"\x00").decode("ascii")
+
+
+
+def setup_relocate_run(wrapper_spec) -> executable.Executable:
+    relocate_exe = Executable(str(wrapper_spec.package.bin_dir() / "relocate.exe"))  # type: ignore
+    # get msvc context from wrapper - needed for finding msvc utils during relocate run
+    relocate_exe.add_default_envmod(
+        spack.user_environment.environment_modifications_for_specs(
+            wrapper_spec, set_package_py_globals=False
+        )
+    )
+    return relocate_exe
+
+
+def bootstrap_relocate() -> executable.Executable:
+    import spack.bootstrap as spack_bootstrap
+
+    with spack_bootstrap.ensure_bootstrap_configuration():
+        # ensure_msvc_relocate_or_raise() may hand back a bare relocate.exe found
+        # via a PATH search, with no MSVC environment attached (see the early
+        # return in ensure_executables_in_path_or_raise). relocate.exe needs the
+        # vcvars-derived INCLUDE/LIB/PATH to find msvc utils (link.exe, lib.exe,
+        # dumpbin.exe, ...), so don't trust its return value: look up the
+        # concrete compiler-wrapper spec ourselves and always attach its
+        # environment, the same way setup_relocate_run does.
+        spack_bootstrap.ensure_msvc_relocate_or_raise()
+        wrapper_spec = next(
+            iter(spack.store.STORE.db.query_local("compiler-wrapper", installed=True)), None
+        )
+        if not wrapper_spec:
+            raise RuntimeError(
+                "Failed to bootstrap the MSVC compiler wrapper: no compiler-wrapper spec "
+                "found in the bootstrap store after bootstrapping relocate.exe"
+            )
+        return setup_relocate_run(wrapper_spec)
+
+
+def relocate(package=None) -> executable.Executable:
+    wrapper_spec = None
+    if package:
+            try:
+                wrapper_spec = package.spec["compiler-wrapper"]
+            except KeyError:
+                pass
+    if not wrapper_spec or not wrapper_spec.installed:
+        # Don't have one associated with our package installed
+        # fine, pull from local db, the functionality we need is
+        # origin agnostic
+        # NOTE: this will need updating if we ever introduce breaking changes
+        # in our relocate behavior
+        wrapper_spec = next(
+            iter(spack.store.STORE.db.query_local("compiler-wrapper", installed=True)), None
+        )
+    if not wrapper_spec:
+        # We need to bootstrap
+        return bootstrap_relocate()
+    return setup_relocate_run(wrapper_spec)
+
+
+@memoized
+def dumpbin(pkg) -> executable.Executable:
+    db_bin_dir = os.path.dirname(pkg["msvc"].cc)
+    dumpbin = executable.which("dumpbin", path=db_bin_dir, required=True)
+    return dumpbin
+
+
+def apply_pe_relocations(
+    pe_targets: Iterable[str],
+    coff_for_target: Dict[str, str],
+    reloc_exe: executable.Executable,
+    ev: EnvironmentModifications,
+    **reloc_kwargs,
+) -> None:
+    """Invoke the compiler wrapper's relocate executable on each PE target (dll or
+    exe). PE files may or may not export symbols (most exes and plugin dlls do not),
+    but references to other PE files inside them still need relocating either way.
+    If ``coff_for_target`` has an import library recorded for a given target, it's
+    passed along via ``--coff`` so the wrapper regenerates its exports to point at
+    the relocated import library.
+    """
+    for pe in pe_targets:
+        args = ["--pe", pe]
+        args.append("--full")
+        if pe in coff_for_target:
+            args.extend(["--coff", coff_for_target[pe]])
+        print(f"relocate args: {args}")
+        reloc_exe(*args, extra_env=ev, **reloc_kwargs)
+
+
+def _win_rpath_import_lib_targets(package, pe_stage_to_prefix: dict) -> Dict[str, str]:
+    """Match each import library under the package prefix against the stage
+    locations recorded in ``pe_stage_to_prefix`` (extracted from each PE's embedded
+    SPACKRESOURCE) and return a mapping from PE path to its associated import
+    library path, for use as the ``--coff`` argument when relocating that PE.
+    """
+    coff_for_target: Dict[str, str] = {}
+    for lib in fs.find(package.spec.prefix, "*.lib"):
+        print(f"processing lib {lib}")
+        if verify_import_lib(lib, package=package):
+            # we have an import lib, determine associated DLL
+            dll_name = get_importlib_target(lib, package=package)
+            if dll_name:
+                print(f"dll name: {dll_name}")
+                prefix_pe = pe_stage_to_prefix.get(dll_name, None)
+                print(f"prefix dll {prefix_pe}")
+                if prefix_pe:
+                    coff_for_target[prefix_pe] = lib
+    return coff_for_target
+
+
+def relocate_win_rpath(package):
+    dlls = fs.find(package.spec.prefix, "*.dll")
+    exes = fs.find(package.spec.prefix, "*.exe")
+    pes = dlls + exes
+    pe_stage_to_prefix = {}
+    # map all PE (dll,exe) prefix locations to the stage
+    for pe in pes:
+        # we don't want to update the rpath to symlinked files
+        if fs.islink(pe):
+            continue
+        print(f"Processing {pe}")
+        # location of PE file at link time (in stage)
+        # is baked into PE file as a resource
+        # extract it
+        stage_pe_loc = extract_spack_id_from_win_pe(pe)
+        if stage_pe_loc:
+            print(f"Discovered stage pe location {stage_pe_loc} for pe {pe}")
+            pe_stage_to_prefix[stage_pe_loc] = pe
+        else:
+            print(f"unable to determine stage pe location for {pe}")
+
+    ev = EnvironmentModifications()
+    ev.set_path("SPACK_RELOCATE_PATH", ["|".join((k, v)) for k, v in pe_stage_to_prefix.items()])
+    ev.set("SPACK_INSTALL_PREFIX", spack.store.STORE.layout.root)
+    ev.set("SPACK_DEBUG_WRAPPER", "ON")
+    coff_for_target = _win_rpath_import_lib_targets(package, pe_stage_to_prefix)
+    apply_pe_relocations(
+        pes, coff_for_target, relocate(package), ev, output=str, error=str, fail_on_error=True
+    )
+
+
+def extract_spack_id_from_win_pe(lib: str) -> Optional[str]:
+    """Extracts the string ID spack of type spackresource from the
+    string table in a given dll
+    Arguments:
+        lib: the dll to extract the string resource from
+    Returns the resource of type SPACKRESOURCE with id spack
+    """
+    if not sys.platform == "win32":
+        return None
+    kernel32 = ctypes.windll.kernel32
+
+    LOAD_LIBRARY_AS_DATAFILE = 0x00000002
+    kernel32.LoadLibraryExW.restype = ctypes.wintypes.HMODULE
+    kernel32.LoadLibraryExW.argtypes = [
+        ctypes.wintypes.LPCWSTR,
+        ctypes.wintypes.HANDLE,
+        ctypes.wintypes.DWORD,
+    ]
+
+    kernel32.FreeLibrary.restype = ctypes.wintypes.BOOL
+    kernel32.FreeLibrary.argtypes = [ctypes.wintypes.HMODULE]
+
+    kernel32.FindResourceW.restype = ctypes.wintypes.HRSRC
+    kernel32.FindResourceW.argtypes = [
+        ctypes.wintypes.HMODULE,
+        ctypes.wintypes.LPCWSTR,
+        ctypes.wintypes.LPCWSTR,
+    ]
+
+    kernel32.LoadResource.restype = ctypes.wintypes.HGLOBAL
+    kernel32.LoadResource.argtypes = [ctypes.wintypes.HMODULE, ctypes.wintypes.HRSRC]
+
+    kernel32.LockResource.restype = ctypes.wintypes.LPVOID
+    kernel32.LockResource.argtypes = [ctypes.wintypes.HGLOBAL]
+
+    kernel32.SizeofResource.restype = ctypes.wintypes.DWORD
+    kernel32.SizeofResource.argtypes = [ctypes.wintypes.HMODULE, ctypes.wintypes.HRSRC]
+
+    RESOURCE_ID = "spack"
+    RESOURCE_TYPE = "SPACKRESOURCE"
+    module_handle = kernel32.LoadLibraryExW(lib, None, LOAD_LIBRARY_AS_DATAFILE)
+    if not module_handle:
+        tty.debug(f"Unable to acquire handle for {lib}: {ctypes.get_last_error()}")
+        kernel32.FreeLibrary(module_handle)
+        return None
+
+    res_info_handle = kernel32.FindResourceW(module_handle, RESOURCE_ID, RESOURCE_TYPE)
+    if not res_info_handle:
+        tty.debug(
+            f"Unable to acquire resource info handle for \
+{lib}:{RESOURCE_ID}:{RESOURCE_TYPE}: {ctypes.get_last_error()}"
+        )
+        kernel32.FreeLibrary(module_handle)
+        return None
+
+    resource_data_handle = kernel32.LoadResource(module_handle, res_info_handle)
+    if not resource_data_handle:
+        tty.debug(
+            f"Unable to acquire resource {res_info_handle} \
+handle for {lib}: {ctypes.get_last_error()}"
+        )
+        kernel32.FreeLibrary(module_handle)
+        return None
+    data_pointer = kernel32.LockResource(resource_data_handle)
+    res_size = kernel32.SizeofResource(module_handle, res_info_handle)
+
+    raw_id_data = ctypes.string_at(data_pointer, res_size)
+    str_id_data = raw_id_data.decode(encoding="utf-8").strip("\x00")  # strip null terminator
+    kernel32.FreeLibrary(module_handle)
+    return str_id_data if res_size else None
+
+
+def get_importlib_target(lib, package=None) -> Optional[str]:
+    reloc = relocate(package)
+    info = reloc("--coff", lib, "--report", output=str)
+    regex = re.compile("DLL: (.*)")
+    match = regex.search(info)
+    if match:
+        pe_name = match.group(1).strip("\r")
+        return pe_name
+
+
+def verify_import_lib(lib: str, package=None) -> bool:
+    relocate_exe = relocate(package)
+    try:
+        relocate_exe("--coff", lib, "--verify", ignore_errors=[1])
+    except executable.ProcessError:
+        tty.debug(f"Cannot verify library {lib} as COFF.")
+    return relocate_exe.returncode == 0
+
+
+def collect_import_exports(pkg, lib):
+    db = dumpbin(pkg)
+    raw_exports = db("/NOLOGO", "/EXPORTS", lib, output=str).split("\n")
+    # first 8 lines are boilerplate and not useful
+    raw_exports = raw_exports[8:]
+    exports = []
+    for export_line in raw_exports:
+        if export_line == "  Summary\r":
+            # exports end just before this section, terminate
+            break
+        sanitized_line = export_line.strip("\r").strip(" ")
+        # some lines define a mangled symbol and their unmangled
+        # method declaration, delinated by a whitespace
+        # we only care about the mangled symbol name to
+        # match with the PE output
+        if " " in sanitized_line:
+            sanitized_line = sanitized_line.split(" ")[0]
+        if not sanitized_line:
+            # there are a couple blank lines, just skip them
+            continue
+        exports.append(sanitized_line)
+    return exports
 
 
 def _buildcache_import_lib_targets(
@@ -121,7 +365,7 @@ def relocate_windows_binaries(
 
     coff_for_target = _buildcache_import_lib_targets(targets, all_prefixes)
     pe_targets = [t for t in targets if t.endswith(".dll") or t.endswith(".exe")]
-    sfs.apply_pe_relocations(pe_targets, coff_for_target, sfs.relocate(spec.package), ev, fail_on_error=True)
+    apply_pe_relocations(pe_targets, coff_for_target, relocate(spec.package), ev, fail_on_error=True)
 
 
 def _macho_find_paths(orig_rpaths, deps, idpath, prefix_to_prefix):
